+++ lib/src/connectivity_service.dart
@@ -1,12 +1,13 @@
 import 'package:connectivity_plus/connectivity_plus.dart';
 
 class ConnectivityService {
-  static Stream<ConnectivityResult> get connectionStream {
+  static Stream<List<ConnectivityResult>> get connectionStream {
     return Connectivity().onConnectivityChanged;
   }
 
   static Future<bool> isOnline() async {
     final result = await Connectivity().checkConnectivity();
+    // ignore: unrelated_type_equality_checks
     return result != ConnectivityResult.none;
   }
 }